Revert to off-site test repository as default about sourcing
To have the change log properly track the android branch, it either
needs to be configured as the default branch, or the api URL has to use
the general project activity format which will lead to extreme excess
parsing and formatting time.

To avoid these issues, the androidui branch is set to default at this
repository and the commits mirror the reicast repo.