Trying out various ideas for loop traces
This uses sampling as a method, and is just an experiment to
test out how many paths / how easy it is to detect hot traces
during the runtime of the emulator.

After first tests
- complex loops 1 level: 25-35%
- complex loops 2 levels: 60-80%
- simple loops 2 levels: 20-25%

Where complex loops include indirect flow.

I need to test out and verify w/ RTS etc