Adds support for NO_VIRTUAL_CFG define. This is required for Launchers that make extensive use of the "-config" command line option. Without it the RTC always resets to default because it parses a virtual config instead of the emu.cfg file.
RetroPie is one of the main usage cases for this mode. If any RetroPie devs try these changes, be sure to open the Makefile and enable NO_VIRTUAL_CONFIG at
the top of the file.

I have rearanged the cfgopen() and parsecommandline() functions to accomodate this.

Add minor fixes for tab/space corrections.
Add minor fix for missing guards around closing evdev controllers.