Proper configuration for triggers / bumpers
This allows the trigger and bumoer buttons to be interchangeable and attempts to address the onscreen buttons being limited to a boolean on / off in the same way Android handles it. There may be the capability to detect degree of pressure on the screen, but implementation will be dependant on the limitations, such as iOS version and hardware.