gl4: use VAOs, use normalized coords in abuffer, fix TR depth testing
In non-autosort, the abuffer depth test should only be done for the second
pixel (if the first writes to the depth buffer). The depth test is already done
in the fragment shader. Fixes flashlight and lightning effects in Alone in the
Dark.
Fix dimensions of buffers if RTT needs a bigger buffer than the screen.
Use normalized coordinates in abuffer.
Don't draw VMUs and crosshair on textures.
Use VAOs everywhere

Tom Clancy's Rainbow 6 needs RTT to vram option