Updates to Makefile
- Make sure we can override HAVE_LTCG, HAVE_CLANG and BOARD (for Odroid)
- Make sure we also pass -flto to the linker when building with HAVE_LTCG

This makes it possible to easily cross-compile like:
make HAVE_CLANG=1 ARCH=arm platform=odroid BOARD=ODROID-XU3

Also when compiling natively on Odroid boards, the /proc/cpuinfo will not list the model (e.g. ODROID-XU3) when you are running a mainline kernel.
That's why it's nice to be able to override the BOARD variable as well.