Fully enabled OpenMP (affects TexCache.cpp)
Since we're already requiring OpenMP support for now to include the refrend and softrends, it makes sense to use OpenMP everywhere we can. In this case, there was just one other place that used it, but a preprocessor directive was preventing it from being used even though it's available.

Later if/when we can remove the OpenMP requirement entirely to allow building with stock Xcode LLVM, we can set TARGET_NO_OPENMP again.