macOS: Added basic console output window
With Mac apps, it's a bit of a pain to view the console output of an app. You have to open the Console application--different from Terminal--and filter by your app name or you have to run the raw binary in Terminal which has other issues like missing images in the UI or missing libraries.

Many people don't know about this, so I added a "Show Console" option that allows you to view the redirected stderr/stdout in a window while still printing stderr/stdout normally (so it's viewable in the Console app or if you run the raw binary from Terminal).