Fixed virtual serial when nothing is connected
When nothing is connected to the tty file, the write buffer gets filled (which is expected), so we needed to add a check for errno != EAGAIN. Also skmp noticed that pty_master was not being initialized to -1 when virtual serial was turned off, also causing problems.

Tested working with previously 100% non-working cdis, and tested virtual tty using screen. Also tested again with dcload-serial, which was working before anyway since it was reading from the tty.