Fixed Werker CI builds by pinning to Ubuntu 18.04
Before it was just specifying "ubuntu" as the docker image to build with, which is the same as using "ubuntu:latest". A few weeks ago, Ubuntu updated their "latest" tag to point to the new 20.04 image rather than the previous 18.04 LTS. This caused a minor change in the installation of the tzdata package (and possibly other stuff).

This commit just pins the docker image to Ubuntu 18.04 so it won't randomly break out from under us any time Ubuntu changes their "latest" image tag.